Azure_Functions/Func_with_flask/Import_pdf/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
flask
azure-core
azure-ai-formrecognizer
azure-cosmos